\name{FindMinIC-package}
\alias{FindMinIC-package}
\docType{package}
\title{
Find Minimum Information Criterion package
}
\description{
Find and rank the best models by information criterion such as AIC.
Looks at models of all possible combinations of the candidate variables with fixed variables always included.
}
\author{
Nicholas Lange, Tom Fletcher, Kristen Zygmunt
}
\references{
\cite{Burnham, K. P.; Anderson, D. R. (2004), "Multimodel inference: understanding AIC and BIC in Model Selection", Sociological Methods and Research 33: 261-304.}
}
\keyword{ package }
\keyword{ models }
\keyword{ multivariate }
\seealso{
\code{\link[FindMinIC]{FindMinIC}}
}

